+++ app/src/main/java/com/example/todoappwithjavagradle/controller/GlobalExceptionHandler.java
@@ -4,9 +4,18 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.servlet.ModelAndView;
 
+/**
+ * アプリケーション全体の例外処理を行う
+ */
 @ControllerAdvice
 public class GlobalExceptionHandler {
 
+    /**
+     * 例外が発生した際の処理を行う
+     * 
+     * @param ex 発生した例外
+     * @return エラーページのModelAndViewオブジェクト
+     */
     @ExceptionHandler(Exception.class)
     public ModelAndView handleException(Exception ex) {
         ModelAndView modelAndView = new ModelAndView();
@@ -14,20 +23,4 @@ public class GlobalExceptionHandler {
         modelAndView.addObject("errorMessage", ex.getMessage());
         return modelAndView;
     }
-
-    // @ExceptionHandler(ItemNotFoundException.class)
-    // public ModelAndView handleItemNotFoundException(ItemNotFoundException ex) {
-    //     ModelAndView modelAndView = new ModelAndView();
-    //     modelAndView.setViewName("error");
-    //     modelAndView.addObject("errorMessage", ex.getMessage());
-    //     return modelAndView;
-    // }
-
-    // @ExceptionHandler(UsernameNotFoundException.class)
-    // public ModelAndView handleUsernameNotFoundException(UsernameNotFoundException ex) {
-    //     ModelAndView modelAndView = new ModelAndView();
-    //     modelAndView.setViewName("login");
-    //     modelAndView.addObject("errorMessage", "ログイン情報が正しくありません。再度試してください。");
-    //     return modelAndView;
-    // }
 }
