+++ app/src/main/java/com/example/todoappwithjavagradle/entity/Item.java
@@ -15,7 +15,7 @@ import lombok.Data;
 
 
 /**
- * ユーザー情報 Entity
+ * アイテム情報エンティティクラス
  */
 @Entity
 @Data
@@ -45,9 +45,19 @@ public class Item {
     private java.sql.Timestamp updatedAt;
 
     public Item() {
-
     }
     
+    /**
+     * コンストラクタ
+     * 
+     * @param id         アイテムID
+     * @param userId     ユーザーID
+     * @param title      タイトル
+     * @param state      状態
+     * @param timeLimit  期限
+     * @param createdAt  作成日時
+     * @param updatedAt  更新日時
+     */
     public Item(Integer id, Integer userId, String title, Integer state, Date timeLimit, Timestamp createdAt, Timestamp updatedAt) {
         this.id = id;
         this.userId = userId;
@@ -58,6 +68,9 @@ public class Item {
         this.updatedAt = updatedAt;
     }
 
+    /**
+     * エンティティが作成されたときの処理
+     */
     @PrePersist
     protected void onCreate() {
         Timestamp ts = new Timestamp(System.currentTimeMillis());
@@ -65,6 +78,9 @@ public class Item {
         this.updatedAt = ts;
     }
 
+    /**
+     * エンティティが更新されたときの処理
+     */
     @PreUpdate
     protected void onUpdate() {
         this.updatedAt = new Timestamp(System.currentTimeMillis());
