+++ app/src/main/java/com/example/todoappwithjavagradle/exception/ItemNotFoundException.java
@@ -1,7 +1,15 @@
 package com.example.todoappwithjavagradle.exception;
 
+/**
+ * アイテムが見つからない場合にスローされる例外
+ */
 public class ItemNotFoundException extends RuntimeException {
 
+    /**
+     * コンストラクタ
+     * 
+     * @param message 例外メッセージ
+     */
     public ItemNotFoundException(String message) {
         super(message);
     }
