+++ app/src/main/java/com/example/todoappwithjavagradle/repository/ItemRepository.java
@@ -8,9 +8,16 @@ import org.springframework.stereotype.Repository;
 import com.example.todoappwithjavagradle.entity.Item;
 
 /**
- * アイテム情報 Repository
+ * アイテム情報にアクセスするためのリポジトリインターフェース
  */
 @Repository
 public interface ItemRepository extends JpaRepository<Item, Integer>  {
+
+    /**
+     * 指定されたユーザーIDに関連するすべてのアイテム情報を取得
+     *
+     * @param userId ユーザーID
+     * @return ユーザーに関連するアイテム情報のリスト
+     */
     List<Item> findByUserId(Integer userId);
 }
