+++ app/src/main/java/com/example/todoappwithjavagradle/service/ItemService.java
@@ -8,28 +8,31 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.example.todoappwithjavagradle.config.AttributeKey;
-import com.example.todoappwithjavagradle.dto.ItemRequest;
+import com.example.todoappwithjavagradle.Form.ItemRequest;
 import com.example.todoappwithjavagradle.entity.Item;
 import com.example.todoappwithjavagradle.repository.ItemRepository;
+import com.example.todoappwithjavagradle.util.AttributeKey;
 
 import jakarta.servlet.http.HttpSession;
 
 /**
- * アイテム情報 Service
+ * アイテム情報を操作するサービスクラス
  */
 @Service
 public class ItemService {
 
-    /**
-    * アイテム情報取得
-    */
     @Autowired
     ItemRepository itemRepository;
 
     @Autowired
     private HttpSession httpSession;
 
+    /**
+     * 特定のユーザーに関連するすべてのアイテム情報を取得します。
+     *
+     * @param userId ユーザーID
+     * @return アイテム情報のリスト
+     */
     public List<Item> searchAll(Integer userId) {
         // アイテムTBLの内容を全検索
         return itemRepository.findByUserId(userId);
@@ -37,13 +40,14 @@ public class ItemService {
 
 
   /**
-   * アイテム情報 新規登録
-   * @param ItemRequest
+   * 新しいアイテム情報を登録
+   *
+   * @param itemRequest 登録するアイテム情報のリクエスト
    */
   public void create(ItemRequest ItemRequest) {
 
     Item item = new Item();
-    Integer userId = (Integer) httpSession.getAttribute(AttributeKey.USER_ID.name());
+    Integer userId = Integer.parseInt(httpSession.getAttribute(AttributeKey.USER_ID.getValue()).toString());
 
     // 日付変換
     SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd");
@@ -61,8 +65,9 @@ public class ItemService {
 
 
   /**
-   * 削除.
-   * @param id
+   * 指定されたIDのアイテム情報を削除
+   *
+   * @param id 削除するアイテムのID
    */
   public void delete(Integer id) {
     if (id == null) {
