+++ app/src/main/java/com/example/todoappwithjavagradle/service/UserService.java
@@ -4,12 +4,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import com.example.todoappwithjavagradle.config.LoginType;
 import com.example.todoappwithjavagradle.entity.User;
 import com.example.todoappwithjavagradle.repository.UserRepository;
+import com.example.todoappwithjavagradle.util.LoginType;
 
 import jakarta.servlet.http.HttpSession;
 
+/**
+ * ユーザーサービスクラス
+ */
 @Service
 public class UserService {
 
@@ -22,7 +25,14 @@ public class UserService {
     @Autowired
     HttpSession httpSession;
 
-
+    /**
+     * プロバイダーからのユーザー登録を処理
+     *
+     * @param oauth2UserId OAuth2ユーザーID
+     * @param username     ユーザー名
+     * @param loginType    ログインタイプ
+     * @return ユーザーID
+     */
     public Integer signupUserFromProvider(String oauth2UserId, String username, String loginType) {
         // GitHub or Google
         User user = new User(oauth2UserId, null, username, loginType);
@@ -32,6 +42,13 @@ public class UserService {
         return userId;
     }
 
+    /**
+     * フォームからのユーザー登録を処理
+     *
+     * @param username ユーザー名
+     * @param password パスワード
+     * @return ユーザーID
+     */
     public Integer signupUserFromForm(String username, String password) {
         // パスワードをハッシュ化して保存
         String encodedPassword = passwordEncoder.encode(password);
@@ -42,6 +59,12 @@ public class UserService {
         return userId;
     }
 
+    /**
+     * ユーザー名からユーザーエンティティを取得
+     *
+     * @param username ユーザー名
+     * @return ユーザーエンティティ
+     */
     public User getUserByUsername(String username) {
         return userRepository.findByUsername(username);
     }
